LayoutTests/platform/blackberry/editing/pasteboard/paste-xml-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 3 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
CONSOLE MESSAGE: line 895: ReferenceError: Can't find variable: editingTest
foo bar baz
